identity server configuration

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -3,6 +3,10 @@ using AuctionService.DtOs;
 using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
+using AutoMapper.QueryableExtensions;
+using Contract;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,93 +18,117 @@ public class AuctionsController : ControllerBase
 {
     private readonly AuctionDbContext _context;
     private readonly IMapper _mapper;
+    private readonly IPublishEndpoint _publishEndpoint;
 
-    public AuctionsController(AuctionDbContext context, IMapper mapper)
-   {
+    public AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
+    {
         _context = context;
         _mapper = mapper;
+        _publishEndpoint = publishEndpoint;
     }
 
     [HttpGet]
-    public async Task<ActionResult<List<AuctionDTo>>> GetAllAuction()
+    public async Task<ActionResult<List<AuctionDTo>>> GetAllAuction(string date)
     {
-         var auctions = await _context.Auctions
+        var query = _context.Auctions.OrderBy(x => x.Item.Make).AsQueryable();
+
+        if (!string.IsNullOrEmpty(date))
+        {
+            query = query.Where(x => x.UpdatedAt.CompareTo(DateTime.Parse(date).ToUniversalTime()) > 0);
+        }
+
+        /*var auctions = await _context.Auctions
              .Include(x => x.Item)
              .OrderBy(x => x.Item.Make)
              .ToListAsync();
         
-        return _mapper.Map<List<AuctionDTo>>(auctions);
-    }
-    
+        return _mapper.Map<List<AuctionDTo>>(auctions);*/
 
-    [HttpGet("{id}", Name="GetAuctionById")]
-    public async Task<ActionResult<AuctionDTo>> GetAuctionById( Guid id)
+        return await query.ProjectTo<AuctionDTo>(_mapper.ConfigurationProvider).ToListAsync();
+    }
+
+
+    [HttpGet("{id}", Name = "GetAuctionById")]
+    public async Task<ActionResult<AuctionDTo>> GetAuctionById(Guid id)
     {
         var auction = await _context.Auctions
           .Include(x => x.Item)
           .FirstOrDefaultAsync(x => x.Id == id);
 
-          if(auction is null) return NotFound();
+        if (auction is null) return NotFound();
 
-         return _mapper.Map<AuctionDTo>(auction);
+        return _mapper.Map<AuctionDTo>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDTo>> CreateAuction(CreateAuctionDTo auctionDTo)
     {
-       var auction = _mapper.Map<Auction>(auctionDTo);
+        var auction = _mapper.Map<Auction>(auctionDTo);
 
-       //todo : add current user as seller
-       auction.Seller = "test"; 
 
-       _context.Auctions.Add(auction);
+        auction.Seller = User.Identity.Name;
 
-       var result = await _context.SaveChangesAsync() > 0;
+        _context.Auctions.Add(auction);
 
-       if(!result) return BadRequest("Could not save changes to the Db");
+        var newAuction = _mapper.Map<AuctionDTo>(auction);
 
-       return CreatedAtRoute(nameof(GetAuctionById), new {auction.Id}, _mapper.Map<AuctionDTo>(auction));
+        await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
+
+        var result = await _context.SaveChangesAsync() > 0;
+
+        if (!result) return BadRequest("Could not save changes to the Db");
+
+        return CreatedAtRoute(nameof(GetAuctionById), new { auction.Id }, newAuction);
     }
 
+
+    [Authorize]
     [HttpPut("{id}")]
-    public async Task<ActionResult> UpdateAuction( Guid id, UpdateAuctionDto updateAuctionDto)
+    public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
         var auction = await _context.Auctions.Include(x => x.Item)
-             .FirstOrDefaultAsync(x => x.Id ==  id);
+             .FirstOrDefaultAsync(x => x.Id == id);
 
-       if(auction is null) return NotFound();
+        if (auction is null) return NotFound();
 
-       //TODO :check seller == username
-       auction.Item.Make = updateAuctionDto.Make ??  auction.Item.Make;
-       auction.Item.Model = updateAuctionDto.Model ??  auction.Item.Model;
-       auction.Item.Color = updateAuctionDto.Color ??  auction.Item.Color;
-       auction.Item.Mileage = updateAuctionDto.Mileage ??  auction.Item.Mileage;
-       auction.Item.Year = updateAuctionDto.Year ??  auction.Item.Year;
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
-       var result = await _context.SaveChangesAsync() > 0;
-       
-       if (result) return Ok();
+        auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
+        auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
+        auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
+        auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
+        auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
 
-       return BadRequest("problem saving changes");
-       
+        await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
+
+        var result = await _context.SaveChangesAsync() > 0;
+
+        if (result) return Ok();
+
+        return BadRequest("problem saving changes");
+
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _context.Auctions.FindAsync(id);
 
-        if(auction is null) return NotFound();
+        if (auction is null) return NotFound();
 
-        //TODO :check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
 
-        var result = await _context.SaveChangesAsync() > 0; 
+        await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
-        if(!result)  return BadRequest("could not update DB");
+        var result = await _context.SaveChangesAsync() > 0;
+
+        if (!result) return BadRequest("could not update DB");
 
         return Ok();
-        
+
     }
 }
